Moved constants and static methods from DefaultNetworkAddressRules back to NetworkAddressRules to revert breaking change
