+++ src/test/java/com/github/tomakehurst/wiremock/common/NetworkAddressRulesTest.java
@@ -25,7 +25,7 @@ public class NetworkAddressRulesTest {
   @Test
   void allowsAddressIncludedAndNotExcluded() {
     NetworkAddressRules rules =
-        DefaultNetworkAddressRules.builder()
+        NetworkAddressRules.builder()
             .allow("10.1.1.1-10.2.1.1")
             .allow("192.168.1.1-192.168.2.1")
             .deny("10.1.2.3")
@@ -41,7 +41,7 @@ public class NetworkAddressRulesTest {
 
   @Test
   void onlyAllowSingleIp() {
-    NetworkAddressRules rules = DefaultNetworkAddressRules.builder().allow("10.1.1.1").build();
+    NetworkAddressRules rules = NetworkAddressRules.builder().allow("10.1.1.1").build();
 
     assertThat(rules.isAllowed("10.1.1.1"), is(true));
     assertThat(rules.isAllowed("10.1.1.0"), is(false));
@@ -50,7 +50,7 @@ public class NetworkAddressRulesTest {
 
   @Test
   void onlyDenySingleIp() {
-    NetworkAddressRules rules = DefaultNetworkAddressRules.builder().deny("10.1.1.1").build();
+    NetworkAddressRules rules = NetworkAddressRules.builder().deny("10.1.1.1").build();
 
     assertThat(rules.isAllowed("10.1.1.1"), is(false));
     assertThat(rules.isAllowed("10.1.1.0"), is(true));
@@ -60,7 +60,7 @@ public class NetworkAddressRulesTest {
   @Test
   void allowAndDenySingleIps() {
     NetworkAddressRules rules =
-        DefaultNetworkAddressRules.builder().deny("10.1.1.1").allow("10.1.1.3").build();
+        NetworkAddressRules.builder().deny("10.1.1.1").allow("10.1.1.3").build();
 
     assertThat(rules.isAllowed("10.1.1.0"), is(false));
     assertThat(rules.isAllowed("10.1.1.1"), is(false));
@@ -72,7 +72,7 @@ public class NetworkAddressRulesTest {
   @Test
   void allowRangeAndDenySingleIp() {
     NetworkAddressRules rules =
-        DefaultNetworkAddressRules.builder().allow("10.1.1.1-10.1.1.3").deny("10.1.1.2").build();
+        NetworkAddressRules.builder().allow("10.1.1.1-10.1.1.3").deny("10.1.1.2").build();
 
     assertThat(rules.isAllowed("10.1.1.0"), is(false));
     assertThat(rules.isAllowed("10.1.1.1"), is(true));
@@ -84,7 +84,7 @@ public class NetworkAddressRulesTest {
   @Test
   void denyRangeAndAllowSingleIp() {
     NetworkAddressRules rules =
-        DefaultNetworkAddressRules.builder().deny("10.1.1.1-10.1.1.3").allow("10.1.1.2").build();
+        NetworkAddressRules.builder().deny("10.1.1.1-10.1.1.3").allow("10.1.1.2").build();
 
     assertThat(rules.isAllowed("10.1.1.0"), is(false));
     assertThat(rules.isAllowed("10.1.1.1"), is(false));
